fixed submitAttempt, ID for Guests, AnswerCorrect(double call)

+++ backend/src/main/java/org/example/backend/Quiz/Services/QuizService.java
@@ -80,6 +80,9 @@ public class QuizService {
     }
 
     public void deleteQuiz(String id) {
+        if (!quizRepo.existsById(id)) {
+            throw new NoSuchElementException("Quiz not found");
+        }
         quizRepo.deleteById(id);
     }
